Move GPUBufferBase class to new Velaptor.OpenGL.Buffers namespace & folder
