+++ Velaptor/OpenGL/Buffers/GPUBufferBase.cs
@@ -2,13 +2,12 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Velaptor.OpenGL
+namespace Velaptor.OpenGL.Buffers
 {
     // ReSharper disable RedundantNameQualifier
     using System;
     using System.Diagnostics.CodeAnalysis;
     using Velaptor.NativeInterop.OpenGL;
-    using Velaptor.OpenGL.Buffers;
     using Velaptor.Reactables.Core;
     using Velaptor.Reactables.ReactableData;
     using NETSizeF = System.Drawing.SizeF;
